Melhorias app desktop

+++ Finances/FinancasDesktop/Entities/Operation.cs
@@ -7,7 +7,7 @@ namespace FinancasDesktop.Entities
         public DateTime? DateTimeOperation { get; set; }
         public string Ativo { get; set; }
         public string TypeOperation { get; set; }
-        public decimal? Amount { get; set; }
+        public int? Amount { get; set; }
         public decimal? Price { get; set; }
         public int? Account { get; set; }
         public decimal? AveragePrice { get; set; }
